+++ projects/ids-enterprise-ng/src/lib/lookup/soho-lookup.component.ts
@@ -2,6 +2,7 @@
 
 import {
   AfterViewInit,
+  AfterViewChecked,
   ChangeDetectionStrategy,
   Component,
   ElementRef,
@@ -25,7 +26,16 @@ import {
   changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [provideControlValueAccessor(SohoLookupComponent)]
 })
-export class SohoLookupComponent extends BaseControlValueAccessor<any> implements AfterViewInit, OnDestroy {
+export class SohoLookupComponent extends BaseControlValueAccessor<any> implements AfterViewInit, AfterViewChecked, OnDestroy {
+  /**
+   * Used to call updated from the afterViewChecked lifecycle event.
+   */
+  private updateRequired: boolean;
+
+  /**
+   * API for interacting with modal and then in turn the buttons on the dialog.
+   */
+  modal: SohoModalStatic;
 
   /**
    * Available Soho Template control settings as Inputs
@@ -33,63 +43,230 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
    */
   @Input() asobject = false; // set to false for backwards compatibility
 
-  // Make sure you bind the context to the function
-  @Input() set beforeShow(value: SohoLookupBeforeShowFunction) {
-    this._options.beforeShow = value;
-  }
-
   /** Grid columns. */
   @Input() columns: SohoDataGridColumn[];
 
-  @Input() set dataset(data: Object[]) {
+  /**
+   * Changes the dataset being used
+   */
+  @Input() public set dataset(data: Object[]) {
+    if (!data) {
+      return;
+    }
+
     this._dataset = data;
-    if (data && this.jQueryElement && this.lookup.grid) {
-      this.lookup.grid.loadData(data);
+    if (data && this.jQueryElement && this.lookup.settings) {
+      this.lookup.settings.options.dataset = data;
+      this.markForUpdate();
     }
   }
-
-  @Input() set editable(value: boolean) {
-    this._options.editable = value;
+  public get dataset() {
+    return this._dataset;
   }
 
-  @Input() set autoWidth(value: boolean) {
-    this._options.autoWidth = value;
+  /** Custom click event; can be used with a modal dialog and custom list component */
+  @Input() public set  click(value: SohoLookupClickFunction) {
+    this.settings.click = value;
+    if (this.lookup) {
+      this.lookup.settings.click = this.settings.click;
+      this.markForUpdate();
+    }
+  }
+  public get click() {
+    return this.settings.click;
   }
 
-  @Input() set delimiter(value: string) {
-    this._options.delimiter = value;
+  /** If a click method is defined, this flexible object can be passed in. */
+  @Input() public set  clickArguments(value: any) {
+    this.settings.clickArguments = value;
+    if (this.lookup) {
+      this.lookup.settings.clickArguments = this.settings.clickArguments;
+      this.markForUpdate();
+    }
+  }
+  public get clickArguments() {
+    return this.settings.clickArguments;
   }
 
-  @Input() set minWidth(value: number) {
-    this._options.minWidth = value;
+  /** Field to return from the array or can be a function. */
+  @Input() public set field(value: string | SohoLookupFieldFunction) {
+    this.settings.field = value;
+    if (this.lookup) {
+      this.lookup.settings.field = this.settings.field;
+      this.markForUpdate();
+    }
+  }
+  public get field() {
+    return this.settings.field;
   }
 
-  @Input() set clearable(value: boolean) {
-    this._options.clearable = value;
+  /** Dialog title or takes the label + Lookup. */
+  @Input() public set title(value: string) {
+    this.settings.title = value;
+    if (this.lookup) {
+      this.lookup.settings.title = this.settings.title;
+      this.markForUpdate();
+    }
+  }
+  public get title() {
+    return this.settings.title;
   }
 
-  @Input() set field(value: string | SohoLookupFieldFunction) {
-    this._options.field = value;
+  /** Pass dialog buttons or Cancel / Apply. */
+  @Input() public set buttons(value: SohoModalButton[]) {
+    this.settings.buttons = value;
+    if (this.lookup) {
+      this.lookup.settings.buttons = this.settings.buttons;
+      this.markForUpdate();
+    }
+  }
+  public get buttons() {
+    return this.settings.buttons;
   }
 
-  @Input() set match(match: SohoDataGridMatchFunction) {
-    this._options.match = match;
+  /** Options to pass to the underlying data grid. */
+  @Input() public set options(value: SohoDataGridOptions) {
+    this.settings.options = value;
+    if (this.lookup) {
+      this.lookup.settings.options = this.settings.options;
+      this.markForUpdate();
+    }
+  }
+  public get options() {
+    return this.settings.options;
   }
 
-  // Make sure you bind the context to the function
-  @Input() set click(value: SohoLookupClickFunction) {
-    this._options.click = value;
+  /**
+   * Used to manage data prior to showing the lookup.
+   *
+   * For example:
+   *  - When the button is clicked, show a loading dialog and make the request for
+   *    lookup grid data.
+   *  - Upon receiving grid data, set lookup.settings.options for the columns and dataset.
+   *  - Then call grid() to build the grid and complete the lookup call.
+   */
+  @Input() public set beforeShow(value: SohoLookupBeforeShowFunction) {
+    this.settings.beforeShow = value;
+    if (this.lookup) {
+      this.lookup.settings.beforeShow = this.settings.beforeShow;
+      this.markForUpdate();
+    }
+  }
+  public get beforeShow() {
+    return this.settings.beforeShow;
   }
 
-  @Input() set title(value: string) {
-    this._options.title = value;
+  /** Custom modal content. */
+  @Input() public set modalContent(value:  JQuery | string) {
+    this.settings.modalContent = value;
+    if (this.lookup) {
+      this.lookup.settings.modalContent = this.settings.modalContent;
+      this.markForUpdate();
+    }
+  }
+  public get modalContent() {
+    return this.settings.modalContent;
+  }
+
+  /** Can the user type random text into the field. */
+  @Input() public set editable(value:  boolean) {
+    this.settings.editable = value;
+    if (this.lookup) {
+      this.lookup.settings.editable = this.settings.editable;
+      this.markForUpdate();
+    }
+  }
+  public get editable() {
+    return this.settings.editable;
+  }
+
+  /** If set to false the dialog wont apply the value on clicking a value. */
+  @Input() public set autoApply(value:  boolean) {
+    this.settings.autoApply = value;
+    if (this.lookup) {
+      this.lookup.settings.autoApply = this.settings.autoApply;
+      this.markForUpdate();
+    }
+  }
+  public get autoApply() {
+    return this.settings.autoApply;
+  }
+
+  /** Function used to match the search term to the data. */
+  @Input() public set match(value: SohoDataGridMatchFunction) {
+    this.settings.match = value;
+    if (this.lookup) {
+      this.lookup.settings.match = this.settings.match;
+      this.markForUpdate();
+    }
+  }
+  public get match() {
+    return this.settings.match;
+  }
+
+  /** A function that fires to let you validate form items on open and select. */
+  @Input() public set validator(value: SohoLookupValidatorFunction) {
+    this.settings.validator = value;
+    if (this.lookup) {
+      this.lookup.settings.validator = this.settings.validator;
+      this.markForUpdate();
+    }
+  }
+  public get validator() {
+    return this.settings.validator;
+  }
+
+  /** Set the width of the input to the width of the selection */
+  @Input() public set autoWidth(value: boolean) {
+    this.settings.autoWidth = value;
+    if (this.lookup) {
+      this.lookup.settings.autoWidth = this.settings.autoWidth;
+      this.markForUpdate();
+    }
+  }
+  public get autoWidth() {
+    return this.settings.autoWidth;
+  }
+
+  /** The character  used to separate data strings */
+  @Input() public set delimiter(value: string) {
+    this.settings.delimiter = value;
+    if (this.lookup) {
+      this.lookup.settings.delimiter = this.settings.delimiter;
+      this.markForUpdate();
+    }
+  }
+  public get delimiter() {
+    return this.settings.delimiter;
+  }
+
+  /** Apply a minimum width to the lookup*/
+  @Input() public set minWidth(value: number) {
+    this.settings.minWidth = value;
+    if (this.lookup) {
+      this.lookup.settings.minWidth = this.settings.minWidth;
+      this.markForUpdate();
+    }
+  }
+  public get minWidth() {
+    return this.settings.minWidth;
+  }
+
+  /**  Add an ability to clear the lookup field with an x */
+  @Input() public set clearable(value: boolean) {
+    this.settings.clearable = value;
+    if (this.lookup) {
+      this.lookup.settings.clearable = this.settings.clearable;
+      this.markForUpdate();
+    }
+  }
+  public get clearable() {
+    return this.settings.clearable;
   }
 
   @Input() multiselect = false;
 
   @Input() name: string;
-
-  @Input() options: SohoDataGridOptions;
 
   // Make sure you bind the context to the function
   @Input() source: SohoDataGridSourceFunction;
@@ -101,9 +278,13 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
    * Should match the Soho event names for the component
    */
   @Output() afteropen: EventEmitter<Object> = new EventEmitter<Object>();
+  @Output() start: EventEmitter<Object> = new EventEmitter<Object>();
+  @Output() complete: EventEmitter<Object> = new EventEmitter<Object>();
   @Output() beforeopen: EventEmitter<Object> = new EventEmitter<Object>();
   @Output() change: EventEmitter<SohoLookupChangeEvent[]> = new EventEmitter<SohoLookupChangeEvent[]>();
+  @Output() input: EventEmitter<Object> = new EventEmitter<Object>();
   @Output() open: EventEmitter<Object> = new EventEmitter<Object>();
+  @Output() close: EventEmitter<Object> = new EventEmitter<Object>();
 
   /**
    * Bind attributes to the host input element
@@ -123,7 +304,7 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
   private lookup: SohoLookupStatic;
 
-  private _options: SohoLookupOptions = {};
+  private settings: SohoLookupOptions = {};
 
   /** Initial dataset */
   private _dataset: Object[];
@@ -163,9 +344,8 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
         source: this.source
       };
 
-      this._options.options = Object.assign(datagridConfig, this.options);
-
-      this.jQueryElement.lookup(this._options);
+      this.settings.options = Object.assign(datagridConfig, this.options);
+      this.jQueryElement.lookup(this.settings);
 
       /**
        * Bind to jQueryElement's events
@@ -175,6 +355,10 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
       this.jQueryElement.on('open', () => this.ngZone.run(() => this.open.emit(null)));
       this.jQueryElement.on('change', (e: any, args: SohoLookupChangeEvent[]) => this.onChange(args));
       this.jQueryElement.on('blur', (e: any) => this.ngZone.run(() => this.touched()));
+      this.jQueryElement.on('start', () => this.ngZone.run(() => this.start.emit(null)));
+      this.jQueryElement.on('complete', () => this.ngZone.run(() => this.complete.emit(null)));
+      this.jQueryElement.on('input', () => this.ngZone.run(() => this.input.emit(null)));
+      this.jQueryElement.on('close', () => this.ngZone.run(() => this.close.emit(null)));
 
       this.lookup = this.jQueryElement.data('lookup');
 
@@ -182,6 +366,88 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
         this.lookup.element.val(this.internalValue);
       }
     });
+  }
+
+  /**
+   * Find the row and select it based on select value / function / field value
+   */
+  public selectRowByValue(field: String, value: String): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.selectRowByValue(field, value));
+    }
+  }
+
+  /** Get the selected rows and return them to the UI **/
+  public insertRows(): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.insertRows());
+    }
+  }
+
+  /** Enable the input. **/
+  public enable(): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.enable());
+    }
+  }
+
+  /** Disable the input. **/
+  public disable(): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.disable());
+    }
+  }
+
+  /** Make the input readonly. **/
+  public readonly(): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.readonly());
+    }
+  }
+
+  /** Input is disabled or not **/
+  public isReadonly(): boolean {
+    if (this.lookup) {
+      return this.ngZone.runOutsideAngular(() => {
+        return this.lookup.isReadonly();
+      });
+    }
+  }
+
+  /**  Send in a new data set to display in the datagrid in the lookup. **/
+  public updateDataset(dataset: Object[], pagerInfo: SohoPagerPagingInfo): void {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.updateDataset(dataset, pagerInfo));
+    }
+  }
+
+  /**
+   * Call to notify the lookup about any dom/settings changes
+   */
+  public updated(settings?: any): void {
+    if (settings) {
+      this.settings = Soho.utils.mergeSettings(this.element[0], settings, this.settings);
+    }
+
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.updated(this.settings));
+    }
+  }
+
+  /**
+   * Destroys the modal lookup.
+   */
+  destroy() {
+    if (this.lookup) {
+      this.ngZone.runOutsideAngular(() => this.lookup.destroy());
+    }
+  }
+
+  ngAfterViewChecked() {
+    if (this.updateRequired) {
+      this.updated();
+      this.updateRequired = false;
+    }
   }
 
   ngOnDestroy() {
@@ -203,6 +469,8 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
   modalOpened(args: any[]) {
     this.ngZone.run(() => {
+      this.modal = this.lookup.modal;
+
       /**
        * Temporary fix for inability for grid to async call data and resize modal on returned
        * values (only necessary when the page size is large enough to make the datagrid larger
@@ -237,13 +505,6 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
       // todo: why is this here, value is not defined anywhere, not sure where it's used so reluctant to remove it.
       event.values = this.internalValue;
-
-      // todo: theo: thinking it should be this instead - Phillip 2018/04/02
-      // if (!event) {
-      //   event = [{}];
-      // }
-      // event[0].value = this.internalValue;
-
       this.change.emit(event);
     });
   }
@@ -275,10 +536,10 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
       let current = '';
 
       if (typeof toProcess[i] === 'object') {
-        if (typeof this._options.field === 'function') {
-          current = (this._options.field as SohoLookupFieldFunction)(toProcess[i], this.lookup.element, this.lookup.grid);
+        if (typeof this.settings.field === 'function') {
+          current = (this.settings.field as SohoLookupFieldFunction)(toProcess[i], this.lookup.element, this.lookup.grid);
         } else {
-          current = (toProcess[i] as any)[this._options.field as string];
+          current = (toProcess[i] as any)[this.settings.field as string];
         }
       } else {
         current = toProcess[i] as string;
@@ -342,17 +603,8 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
       this.internalValue = event.map(val => this.asobject !== false ? val.data : this.processValue(val.data));
     }
   }
-}
 
-declare abstract class OnBeforeLookupShow {
-  /**
-   * Used to manage data prior to showing the lookup.
-   *
-   * For example:
-   *  - When the button is clicked, show a loading dialog and make the request for
-   *    lookup grid data.
-   *  - Upon receiving grid data, set lookup.settings.options for the columns and dataset.
-   *  - Then call grid() to build the grid and complete the lookup call.
-   */
-  abstract onBeforeLookupShow: (lookup: any, grid: (gridOptions: Object) => {}) => any;
+  private markForUpdate(): void {
+    this.updateRequired = true;
+  }
 }