+++ src/app/lookup/lookup.demo.ts
@@ -1,4 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { SohoLookupComponent } from 'ids-enterprise-ng';
+import { TemplateService } from './template.service';
+import { Asset } from './asset';
 
 import {
   checkboxColumn,
@@ -16,11 +19,32 @@ export interface FakeResponse {
   selector: 'app-lookup-demo',
   templateUrl: 'lookup.demo.html',
 })
-export class LookupDemoComponent {
+export class LookupDemoComponent implements OnInit {
+  @ViewChild('templateId', { static: true }) sohoLookupComponent: SohoLookupComponent;
+  @ViewChild('toggleButtons', { static: true }) sohoLookupRef: SohoLookupComponent;
+
   public columns_product: SohoDataGridColumn[];
   public columns_multi: SohoDataGridColumn[];
-
+  public entityIds: string;
   public data_product: any[];
+  public customButtons: SohoModalButton[] = [
+    {
+      text: 'Enable', click: () => {
+        const api = this.sohoLookupRef.modal.buttonsetAPI;
+        api.at(2).disabled = false;
+        api.at(3).disabled = false;
+      }
+    },
+    {
+      text: 'Disable', click: () => {
+        const api = this.sohoLookupRef.modal.buttonsetAPI;
+        api.at(2).disabled = true;
+        api.at(3).disabled = true;
+      }
+    },
+    { text: 'Cancel', click: () => { this.sohoLookupRef.modal.close(); } },
+    { text: 'Submit', click: () => { this.sohoLookupRef.modal.close(); }, isDefault: true }
+  ];
   public model: any = {
     single: null,
     singleexists: '1212121',
@@ -64,12 +88,72 @@ export class LookupDemoComponent {
     asyncexists: '2342203',
   };
   public showModel = false;
+  public templates: Array<Asset>;
 
   // So we can bind 'this' to the source function passed to the lookup control
   public context = this;
 
-  constructor() {
+  constructor(private service: TemplateService) {
     this.setupProducts();
+  }
+
+  ngOnInit(): void {
+    const lookupOptions = {
+      columns: this.getAssetColumns(),
+      sortable: false,
+      alternateRowShading: false,
+      selectable: 'multiple',
+      clickToSelect: true,
+      paging: true,
+      toolbar: {
+        results: true,
+        dateFilter: false,
+        keywordFilter: true,
+        advancedFilter: false,
+        actions: false,
+        views: true,
+        rowHeight: true,
+        collapsibleFilter: false,
+        fullWidth: false,
+      }
+    } as SohoDataGridOptions;
+    this.sohoLookupComponent.options = lookupOptions;
+    this.service.getAvailableTemplates().subscribe(result => {
+      this.templates = result.data;
+    });
+  }
+
+  private getAssetColumns(): Array<SohoDataGridColumn> {
+    const columns = [
+      {
+        id: 'selectionCheckbox',
+        sortable: false,
+        formatter: Soho.Formatters.SelectionCheckbox,
+        align: 'center'
+      },
+      {
+        id: 'templateId',
+        sortable: false,
+        field: 'templateId'
+      },
+      {
+        id: 'templateDisplayName',
+        name: 'Template',
+        field: 'displayName',
+        sortable: false
+      }];
+    const sohoGridColumns: Array<SohoDataGridColumn> = [];
+    columns.forEach(column => {
+      const sohoGridColumn = {
+        name: column.name,
+        field: column.field,
+        formatter: column.formatter,
+        sortable: column.sortable,
+        id: column.id,
+      } as SohoDataGridColumn;
+      sohoGridColumns.push(sohoGridColumn);
+    });
+    return sohoGridColumns;
   }
 
   requestData(filter?: string, page?: number, pagesize?: number): Promise<FakeResponse> {
@@ -142,8 +226,8 @@ export class LookupDemoComponent {
   }
 
   // Example of calling before show to cancel the opening by returning false
-  onBeforeShow = (api, response) => {
-    console.log(this);
+  onBeforeShow = (api: any, response: any) => {
+    console.log(api, response);
     // 1. Do something like an ajax call.
 
     // 2. if no rows and you dont want to open
@@ -154,27 +238,27 @@ export class LookupDemoComponent {
     // api.settings.options.columns = data[0].columns;
     // api.settings.options.dataset = data[0].dataset;
     // 5. And the response which will open the dialog
-    response([]);
+    response();
   }
 
   // Example of custom displaying
-  onField = (row, field, grid) => {
+  onField = (row: any, field: any, grid: any) => {
     console.log(row, field, grid, this);
     return row.productId + '|' + row.productName;
   }
 
   // Example of custom matching
-  onMatch(value, row, field, grid) {
+  onMatch(value: any, row: any, field: any, grid: any) {
     console.log(row, field, grid);
     return ((row.productId + '|' + row.productName) === value);
   }
 
   onDirty(event: SohoTrackDirtyEvent) {
-    console.log('lookup.onDirty');
+    console.log('lookup.onDirty', event);
   }
 
   onPristine(event: SohoTrackDirtyEvent) {
-    console.log('lookup.onPristine');
+    console.log('lookup.onPristine', event);
   }
 
   onChange(event: any) {